src/integrations/gmail/requirements.txt
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
mcp>=0.9.0
